(fix): connection manager evict stale connections (#7924)
* Eviction config in connection-manager.js 

Added configuration property `evictionRunIntervalMillis` to node-pool instantiation to allow evictions to run against stale connections in the pool. The default value is set to 0 which prevents evictions all together. The added configuration will cause evictions to run every 60 seconds. The setting has been exposed in the Sequelize configuration options.

* Update default pool options in usage.md

Corrected the existing default pooling options in `usage.md` and added new option for evict which controls the interval in which evictions run.

* Added evictionTimeoutMillis to remaining areas

Adding `evictionTimeoutMillis` to remaining configuration areas within `connection-manager.js`

* Adding documentation for `options.pool.evict`

Updating the comment which documents the configuration parameters available when constructing a new _Sequelize_ object.

* Change "between" -> "for" 

Updating the grammar for simplify the understanding of this option.